+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - launch.json: "disassemblerArgs"/"numberOfLines": New parameter to control the number of displayed disassembled lines.
 - Byte registers IXL, IXH, IYL and IYH now show additionally under "Registers 2"
 - Bugfix for a hang when Disassembly should wrap around 0xFFFF.
+- Fixed an "Unverified breakpoint" issue on windows.
 
 # 1.4.9
 - Fixed issue #29: Zsim: load instruction not executed properly for addresses between 0x0000 and 0x3FFF