+++ CHANGELOG.md
@@ -2,14 +2,17 @@
 
 # 1.5.0
 - New architecture for parsing list files. This will make it easier to add parsing for a new assembler.
-- Related changes to launch.json:
+- Changes to launch.json:
 	- "listFiles" removed.
 	- Instead a configuration for each assembler: "sjasmplus", "z80asm" and "z88dk" added with overworked parameters.
+	- "excludeFiles" parameter added to allow exclude certain files from association with execution addresses.
+	- "filter" regex removed.
+	- "disassemblerArgs"/"numberOfLines": New parameter to control the number of displayed disassembled lines.
 - z88dk:
 	- Renamed "z88dkMapFile" to "mapFile".
 	- "addOffset" removed for z88dk. Use "mapFile" instead.
 	- "mapFile" is mandatory now.
-- launch.json: "disassemblerArgs"/"numberOfLines": New parameter to control the number of displayed disassembled lines.
+- launch.json:
 - Byte registers IXL, IXH, IYL and IYH now show additionally under "Registers 2"
 - Bugfix for a hang when Disassembly should wrap around 0xFFFF.
 - Fixed an "Unverified breakpoint" issue on windows.